+++ src/chapters/01/test/ex02.test.ts
@@ -1,0 +1,7 @@
+import { ex02 } from '../ex02'
+
+describe('02. "shoe" + "cold" = "schooled"', () => {
+  it('the letters in "shoe" and "cold" one after the other from head to tail', () => {
+    expect(ex02()).toBe('schooled')
+  })
+})