+++ src/chapters/01/test/ex03.test.ts
@@ -1,0 +1,7 @@
+import { ex03 } from '../ex03'
+
+describe('03. Pi', () => {
+  it('a list whose element presents the number of alphabetical letters in the corresponding word', () => {
+    expect(ex03()).toEqual([3, 1, 4, 1, 5, 9, 2, 6, 5, 3, 5, 8, 9, 7, 9])
+  })
+})