+++ CHANGELOG.md
@@ -1,7 +1,50 @@
 # Changelog
 
+<a name="0.1.0"></a>
+## [v0.1.0](https://github.com/neutron0831/nlp100-ts/compare/v0.0.1...v0.1.0) (2023-03-19)
+
+### Added
+
+- ✨ Add test workflow [[b065843](https://github.com/neutron0831/nlp100-ts/commit/b065843)]
+- ✅ Add Exercise 09 test ([#9](https://github.com/neutron0831/nlp100-ts/issues/9)) [[ef72616](https://github.com/neutron0831/nlp100-ts/commit/ef72616)]
+- ✨ Add Exercise 09 ([#9](https://github.com/neutron0831/nlp100-ts/issues/9)) [[22e1909](https://github.com/neutron0831/nlp100-ts/commit/22e1909)]
+- ✅ Add Exercise 08 test ([#8](https://github.com/neutron0831/nlp100-ts/issues/8)) [[bb3d8c7](https://github.com/neutron0831/nlp100-ts/commit/bb3d8c7)]
+- ✨ Add Exercise 08 ([#8](https://github.com/neutron0831/nlp100-ts/issues/8)) [[5cd6275](https://github.com/neutron0831/nlp100-ts/commit/5cd6275)]
+- ✅ Add Exercise 07 test ([#7](https://github.com/neutron0831/nlp100-ts/issues/7)) [[0c4e222](https://github.com/neutron0831/nlp100-ts/commit/0c4e222)]
+- ✨ Add Exercise 07 ([#7](https://github.com/neutron0831/nlp100-ts/issues/7)) [[2848116](https://github.com/neutron0831/nlp100-ts/commit/2848116)]
+- ✅ Add Exercise 06 test ([#6](https://github.com/neutron0831/nlp100-ts/issues/6)) [[e64af26](https://github.com/neutron0831/nlp100-ts/commit/e64af26)]
+- ✨ Add Exercise 06 ([#6](https://github.com/neutron0831/nlp100-ts/issues/6)) [[c547ed5](https://github.com/neutron0831/nlp100-ts/commit/c547ed5)]
+- ✅ Add Exercise 05 test ([#5](https://github.com/neutron0831/nlp100-ts/issues/5)) [[88670b1](https://github.com/neutron0831/nlp100-ts/commit/88670b1)]
+- ✨ Add Exercise 05 ([#5](https://github.com/neutron0831/nlp100-ts/issues/5)) [[eb83460](https://github.com/neutron0831/nlp100-ts/commit/eb83460)]
+- ✅ Add Exercise 04 test ([#4](https://github.com/neutron0831/nlp100-ts/issues/4)) [[4d98c34](https://github.com/neutron0831/nlp100-ts/commit/4d98c34)]
+- ✨ Add Exercise 04 ([#4](https://github.com/neutron0831/nlp100-ts/issues/4)) [[5b8c8f9](https://github.com/neutron0831/nlp100-ts/commit/5b8c8f9)]
+- ✅ Add Exercise 03 test ([#3](https://github.com/neutron0831/nlp100-ts/issues/3)) [[adf98d9](https://github.com/neutron0831/nlp100-ts/commit/adf98d9)]
+- ✨ Add Exercise 03 ([#3](https://github.com/neutron0831/nlp100-ts/issues/3)) [[89a814a](https://github.com/neutron0831/nlp100-ts/commit/89a814a)]
+- ✅ Add Exercise 02 test ([#2](https://github.com/neutron0831/nlp100-ts/issues/2)) [[716bf77](https://github.com/neutron0831/nlp100-ts/commit/716bf77)]
+- ✨ Add Exercise 02 ([#2](https://github.com/neutron0831/nlp100-ts/issues/2)) [[457663a](https://github.com/neutron0831/nlp100-ts/commit/457663a)]
+- ✅ Add Exercise 01 test ([#1](https://github.com/neutron0831/nlp100-ts/issues/1)) [[f2e4805](https://github.com/neutron0831/nlp100-ts/commit/f2e4805)]
+- ✨ Add Exercise 01 ([#1](https://github.com/neutron0831/nlp100-ts/issues/1)) [[6b71cf5](https://github.com/neutron0831/nlp100-ts/commit/6b71cf5)]
+- ✅ Add Exercise 00 test ([#100](https://github.com/neutron0831/nlp100-ts/issues/100)) [[2676aba](https://github.com/neutron0831/nlp100-ts/commit/2676aba)]
+- ➕ Add Vitest [[cda55ca](https://github.com/neutron0831/nlp100-ts/commit/cda55ca)]
+- ✨ Add Exercise 00 ([#100](https://github.com/neutron0831/nlp100-ts/issues/100)) [[61a8d18](https://github.com/neutron0831/nlp100-ts/commit/61a8d18)]
+
+### Changed
+
+- 💄 Change `dot-color` of `VTimelineItem` to `primary` [[98f988f](https://github.com/neutron0831/nlp100-ts/commit/98f988f)]
+
+### Fixed
+
+- 💚 Fix a build error [[c929b14](https://github.com/neutron0831/nlp100-ts/commit/c929b14)]
+
+### Miscellaneous
+
+- 📝 Add `README.md` of Chapter 1 [[25a8de7](https://github.com/neutron0831/nlp100-ts/commit/25a8de7)]
+- 📝 Add `README.md` of chapters [[525ad90](https://github.com/neutron0831/nlp100-ts/commit/525ad90)]
+- 🩹 Fix the previous tag name in changelog script [[514d3f4](https://github.com/neutron0831/nlp100-ts/commit/514d3f4)]
+
+
 <a name="0.0.1"></a>
-## [v0.0.1](https://github.com/neutron0831/nlp100-ts/compare/v0.0.0...0.0.1) (2023-03-12)
+## [v0.0.1](https://github.com/neutron0831/nlp100-ts/compare/v0.0.0...v0.0.1) (2023-03-12)
 
 ### Added